requirements.txt
Pillow==10.1.0
django==5.0
django-ckeditor==6.7.0